feat(reader): rotate without crash
Before: `parser` and `pageNum` are fields in `ComicPageViewerFragment`,
passed in directly via its constructor. Upon rotation, the fragment is recreated
by the empty constructor, which is not implemented as a parameterized constructor
(for the above 2 variables) exists.
After: `parser` and `pageNum` is now stored in `ComicPageViewerViewModel`.

Possible bug: handling of caching file is now much more difficult, as each
`ComicPageViewerViewModel` has its own parser. Currently, the code to delete
cache file in `ComicParser::close` is commented out.