wip: fix broken hierarchy of `SearchableMetadataDao`
A massive fuck up for an undergrad student who claims to have some good command of Java.

Previously, the interface `SearchableMetadataDao` is not parameterized, and its function
is declared and implemented as follow:

- `SearchableMetadata`:

    ```kotlin
    fun search(name: String, limit: Int = Int.MAX_VALUE): List<SearchableMetadata>
    ```

- `CharacterDao`:

    ```kotlin
    override fun search(name: String, limit: Int): List<Character>
    ```

However, `List<Character>` is not immediately a child of `List<SearchableMetadata>`.
Then `SearchableMetadata` must be parameterized, as done in this commit.